rearrange plotutils
